启用、禁用员工账号代码实现(优化sql提高代码复用性)

+++ sky-server/src/main/java/com/sky/controller/admin/EmployeeController.java
@@ -100,9 +100,15 @@ public class EmployeeController {
         return Result.success(pageResult);
     }
 
+    /**
+     * 启用、禁用员工账号
+     * @param status
+     * @param id
+     * @return
+     */
     @PostMapping("status/{status}")
     @ApiOperation("启用、禁用员工账号")
-    public Result<String> startOrStop(@PathVariable Integer  status,Integer id){
+    public Result<String> startOrStop(@PathVariable Integer  status,Long id){
         log.info("接收到id:{}员工的状态修改请求{}",status,id);
         employeeService.startOrStop(status,id);
         return Result.success();
